# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI Operational POS API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-UNLICENSED
  description: Phase 0 foundation contract with the backend voice transcription slice. Product, auth, sales, inventory, and intent handlers are not declared until implemented.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
...
  /api/v1/categories:
    get:
      operationId: listCategories
      summary: List active product categories for the tenant
      responses:
        '200':
          description: Tenant category list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCategoryList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createCategory
      summary: Create a product category
      description: Category names are unique per tenant (case-insensitive).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCategoryCreated'
        '400':
          description: Category name missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Category name already exists for the tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional machine-readable context for the error.
...
    Category:
      type: object
      required: [id, tenant_id, name, is_active]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    SuccessCategoryList:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        request_id:
          type: string
    SuccessCategoryCreated:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Category'
        request_id:
          type: string
...